Initial Add User DB Code (Still not fully working)

+++ datawiz-api/src/main/java/com/groupten/datawiz/model/User.java
@@ -10,7 +10,7 @@ public class User {
     @Id
     @Column(name = "user_id")
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long userId;
+    private int userId;
 
     private String username;
 
@@ -31,11 +31,11 @@ public class User {
         this.password = password;
     }
 
-    public long getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
